IPRT: socket / address resolving fixes.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@39801 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/posix/process-posix.cpp
@@ -130,6 +130,9 @@ RTR3DECL(int)   RTProcWaitNoResume(RTPROCESS Process, unsigned fFlags, PRTPROCST
 
 RTR3DECL(int) RTProcTerminate(RTPROCESS Process)
 {
+    if (Process == NIL_RTPROCESS)
+        return VINF_SUCCESS;
+
     if (!kill(Process, SIGKILL))
         return VINF_SUCCESS;
     return RTErrConvertFromErrno(errno);
